backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and LLM
openai==1.54.0
langdetect==1.0.9

# Report Generation
openpyxl==3.1.2
reportlab==4.0.9
Pillow==10.3.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.27.2
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Parsing and Analysis
pyyaml==6.0.1
markdown==3.5.2
beautifulsoup4==4.12.3

# Date/Time
python-dateutil==2.8.2
